+++ CHANGELOG.md
@@ -6,8 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Implementation of `FloatDiff` for arrays of size 0 to 32 (inclusive) where the
-  type allows it.
+- Implementation of `FloatEq` and `FloatDiff` for arrays of size 0 to 32
+  (inclusive) where the type allows it.
 
 ## [0.1.3] - 2020-04-07
 ### Added